add footer everywhere
