+++ src/shared/components/layouts/SingleColumnLayout.tsx
@@ -1,14 +1,10 @@
-import { lazy, FC, Suspense } from 'react';
+import { FC, Suspense } from 'react';
 import { Loader } from 'franklin-sites';
 import cn from 'classnames';
 
 import ErrorBoundary from '../error-component/ErrorBoundary';
 
 import './styles/single-column-layout.scss';
-
-const UniProtFooter = lazy(
-  () => import(/* webpackChunkName: "footer" */ './UniProtFooter')
-);
 
 type SingleColumnLayoutProps = {
   className?: string;
@@ -24,11 +20,6 @@ const SingleColumnLayout: FC<SingleColumnLayoutProps> = ({
         <ErrorBoundary>{children}</ErrorBoundary>
       </Suspense>
     </main>
-    <Suspense fallback={null}>
-      <ErrorBoundary>
-        <UniProtFooter className="single-column-layout__footer" />
-      </ErrorBoundary>
-    </Suspense>
   </div>
 );
