+++ src/shared/components/layouts/styles/single-column-layout.scss
@@ -1,15 +1,10 @@
 .single-column-layout {
   min-height: 100%;
-  width: 100vw;
   display: flex;
   flex-direction: column;
   align-items: center;
 
-  & > .single-column-layout__main {
-    flex: 1 0;
-  }
-
-  & > *:not(.single-column-layout__footer) {
+  & > * {
     width: 80vw;
     max-width: 160ch;
   }
